Correctly parse strings
The issue was that we were using the decoded, UTF-16 length as the size
of the buffer to store the encoded string, which means we were sometimes
missing a few bytes, which in turns meant we could not always properly
decode non-ASCII strings. By using the correct length we can directly
call String::from_utf8() without issues.